+++ src/main/java/org/owasp/webgoat/lessons/clientsidefiltering/ShopEndpoint.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.client_side_filtering;
+package org.owasp.webgoat.lessons.clientsidefiltering;
 
 import com.google.common.collect.Lists;
 import lombok.AllArgsConstructor;
