+++ src/main/resources/lessons/clientsidefiltering/css/clientSideFiltering-stage1.css
