+++ src/main/resources/lessons/clientsidefiltering/css/clientSideFilteringFree.css
