+++ src/main/resources/lessons/clientsidefiltering/documentation/ClientSideFiltering_assignment.adoc
