+++ src/main/resources/lessons/clientsidefiltering/documentation/ClientSideFiltering_final.adoc
