+++ src/main/resources/lessons/clientsidefiltering/documentation/ClientSideFiltering_plan.adoc
