+++ src/main/resources/lessons/clientsidefiltering/js/clientSideFiltering.js
