+++ src/main/resources/lessons/clientsidefiltering/js/clientSideFilteringFree.js
