+++ src/main/resources/lessons/clientsidefiltering/lessonSolutions/en/ClientSideFiltering.html
