+++ src/test/java/org/owasp/webgoat/lessons/clientsidefiltering/ClientSideFilteringFreeAssignmentTest.java
@@ -1,25 +1,21 @@
-package org.owasp.webgoat.lessons.client_side_filtering;
+package org.owasp.webgoat.lessons.clientsidefiltering;
 
 import org.hamcrest.CoreMatchers;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.owasp.webgoat.container.plugins.LessonTest;
-import org.owasp.webgoat.lessons.client_side_filtering.ClientSideFiltering;
+import org.owasp.webgoat.lessons.clientsidefiltering.ClientSideFiltering;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import static org.mockito.Mockito.when;
-import static org.owasp.webgoat.lessons.client_side_filtering.ClientSideFilteringFreeAssignment.SUPER_COUPON_CODE;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
-/**
- * @author nbaars
- * @since 5/2/17.
- */
-public class ClientSideFilteringAssignmentTest extends LessonTest {
+public class ClientSideFilteringFreeAssignmentTest extends LessonTest {
 
     @BeforeEach
     public void setup() {
@@ -29,16 +25,23 @@ public class ClientSideFilteringAssignmentTest extends LessonTest {
 
     @Test
     public void success() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.post("/clientSideFiltering/getItForFree")
-                .param("checkoutCode", SUPER_COUPON_CODE))
+        mockMvc.perform(MockMvcRequestBuilders.post("/clientSideFiltering/attack1")
+                .param("answer", "450000"))
                 .andExpect(jsonPath("$.lessonCompleted", CoreMatchers.is(true)));
     }
 
     @Test
-    public void wrongCouponCode() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.post("/clientSideFiltering/getItForFree")
-                .param("checkoutCode", "test"))
-                .andExpect(jsonPath("$.feedback", CoreMatchers.is(messages.getMessage("assignment.not.solved"))))
+    public void wrongSalary() throws Exception {
+        mockMvc.perform(MockMvcRequestBuilders.post("/clientSideFiltering/attack1")
+                .param("answer", "10000"))
+                .andExpect(jsonPath("$.feedback", CoreMatchers.is("This is not the salary from Neville Bartholomew...")))
                 .andExpect(jsonPath("$.lessonCompleted", CoreMatchers.is(false)));
+    }
+
+    @Test
+    public void getSalaries() throws Exception {
+        mockMvc.perform(MockMvcRequestBuilders.get("/clientSideFiltering/salaries"))
+                .andExpect(jsonPath("$[0]", Matchers.hasKey("UserID")))
+                .andExpect(jsonPath("$.length()", CoreMatchers.is(12)));
     }
 }
